numerous updates! :D
git-svn-id: http://www.observium.org/svn/observer/trunk@440 61d68cd4-352d-0410-923a-c4978735b2b8

+++ process-syslog.php
@@ -3,6 +3,11 @@
 
 include("config.php");
 
+if(!$config['enable_syslog']) {
+  echo("Syslog support disabled.\n");
+  exit();
+}
+
 include("includes/syslog.php");
 
 mysql_query("DELETE FROM `syslog` WHERE `processed` = '0' AND `msg` LIKE '%last message repeated%'");
@@ -12,10 +17,11 @@ mysql_query("DELETE FROM `syslog` WHERE `processed` = '0' AND `msg` LIKE '%PM-3-
 mysql_query("DELETE FROM `syslog` WHERE `processed` = '0' AND `msg` LIKE '%RHWatchdog%'");
 mysql_query("DELETE FROM `syslog` WHERE `processed` = '0' AND `msg` LIKE '%Hardware Monitoring%'");
 
-if(!$config['enable_syslog']) { 
-  echo("Syslog support disabled.\n"); 
-  exit(); 
-}
+#mysql_query("DELETE FROM `syslog` WHERE `program` LIKE 'SNMP-3-AUTHFAIL'");
+#mysql_query("DELETE FROM `syslog` WHERE `program` LIKE 'SW_MATM-4-MACFLAP_NOTIF'");
+
+mysql_query("DELETE FROM `syslog` WHERE `priority` = 'debug'");
+
 
 ## Delete all the old old old syslogs (as per config.php variable)
